// src/component/product/types.tsx
export type DetailInfoFormValue = {
  graphicDiameter: number
  basecurve: number
  color: string
  colorCode: string
  material: string
  detailPrice: number
  moisture: number
  productDetailsStock: number
  isSale: number
  detailsExposure: number
  period: number
  imageUrl: string[]
  explanationImageUrl: string[]
  degree: number[]
}